atualizações de rotas
