+++ modulo4/reforco-star-wars/src/App.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import Header from './components/Header';
-import Footer from './components/Footer'; 
-import './App.css';
 import Routes from './Routes';
 
 
@@ -15,7 +13,7 @@ function App() {
 
       <h1>Star Wars</h1>
 
-      <Footer Volte Sempre/>
+      
       </BrowserRouter>
       
   );
