+++ modulo4/reforco-star-wars/src/CharacterListPage/CharacterListPage.js
@@ -1,2 +1,16 @@
 import React from "react";
 import axios from './axios';
+import { ContainerPage, TitlePage } from "../..components/Main";
+
+const Page = () => {
+    return (
+
+        <ContainerPage>
+              <TitlePage>Seja Bem Vindo!</TitlePage>
+        </ContainerPage>
+
+        
+    );
+}
+
+export default Page;
