--- modulo4/reforco-star-wars/src/Pages/Config/index.js
@@ -1,16 +1,0 @@
-import React from "react";
-
-import { ContainerPage, TitlePage } from "../..components/Main";
-
-const Page = () => {
-    return (
-
-        <ContainerPage>
-              <TitlePage>Configurações</TitlePage>
-        </ContainerPage>
-
-        
-    );
-}
-
-export default Page;
