+++ modulo4/reforco-star-wars/src/Routes.js
@@ -1,24 +1,24 @@
 import React from "react";
 import { Routes, Switch } from "react-router-dom";
-import Home from './Pages/Home';
-import Config from './Pages/Config';
+import CharacterDetail from './src/CharacterDetailPage/CharacterDetailPage';
+import CharacterList from './src/CharacterListPage/CharacterListPage';
 
 export default () => {
 
     return(
         <Switch>
-            <Routes exact path="/">
+            {/* <Routes exact path="/">
                 <Home />
 
+            </Routes> */}
+
+            <Routes exact path="/CharacterDetail">
+                <CharacterDetail />
+
             </Routes>
 
-            <Routes exact path="/config">
-                <config />
-
-            </Routes>
-
-            <Routes exact path="/Pagina3">
-                <Pagina3 />
+            <Routes exact path="/CharacterList">
+                <CharacterList />
 
             </Routes>
         </Switch>
